Change ivoiceValue type from uint8_t to uint16_t
Data input for ivoice is 10bits wide, need at least a uint16 buffer to store it.

+++ src/cartridge.c
@@ -51,7 +51,7 @@
    uint8_t ivoiceRead = 0;
    volatile uint8_t ivoiceWrite = 0;
    volatile uint8_t ivoiceRegister[IVOICE_BUF_SIZE] = {0};
-   volatile uint8_t ivoiceValue[IVOICE_BUF_SIZE] = {0};
+   volatile uint16_t ivoiceValue[IVOICE_BUF_SIZE] = {0};
 #endif
 
 extern Cartridge cart;     // main data structure for cart emulation
